% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glinter_net.R
\name{glinter_net}
\alias{glinter_net}
\title{Wrapper for glinteret}
\usage{
glinter_net(x, y, penalty, ...)
}
\arguments{
\item{x}{A matrix or data frame of predictors.}

\item{y}{A vector (numeric or factor)}

\item{penalty}{A sequence of penalty values.}
}
\description{
Wrapper for glinteret
}
\examples{
library(modeldata)
library(dplyr)

ames <-
  ames \%>\%
  select(Sale_Price, Bldg_Type, Neighborhood, Year_Built, Gr_Liv_Area, Full_Bath,
         Half_Bath, Year_Sold, Lot_Area, Central_Air, Longitude, Latitude) \%>\%
  mutate(
    Sale_Price = log10(Sale_Price),
    Baths = Full_Bath  + Half_Bath / 2
  ) \%>\%
  select(-Half_Bath, -Full_Bath)

glinter_net(ames[, -1], ames$Sale_Price)
}
